requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Request handling
python-multipart==0.0.6
pydantic[email]==2.5.0

# PDF processing
pymupdf==1.23.8

# NLP
spacy==3.7.2

# Database
supabase==2.0.3

# ML/Data processing
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.0
